removing unnecessary code
