+++ packages/bridge-ui/src/utils/selectChain.spec.ts
@@ -1,8 +1,7 @@
-import { switchNetwork } from '@wagmi/core';
-import { ethers } from 'ethers';
+import { fetchSigner, switchNetwork } from '@wagmi/core';
+import type { ethers } from 'ethers';
 
 import { mainnetChain, taikoChain } from '../chain/chains';
-import { destChain, srcChain } from '../store/chain';
 import { signer } from '../store/signer';
 import { selectChain } from './selectChain';
 
@@ -10,30 +9,7 @@ jest.mock('../constants/envVars');
 
 jest.mock('@wagmi/core', () => ({
   switchNetwork: jest.fn(),
-}));
-
-jest.mock('ethers', () => {
-  const Web3Provider = jest.fn();
-  Web3Provider.prototype = {
-    getSigner: jest.fn(),
-    send: jest.fn(),
-  };
-  return {
-    ethers: {
-      providers: {
-        Web3Provider,
-      },
-    },
-  };
-});
-
-jest.mock('../store/chain', () => ({
-  srcChain: {
-    set: jest.fn(),
-  },
-  destChain: {
-    set: jest.fn(),
-  },
+  fetchSigner: jest.fn(),
 }));
 
 jest.mock('../store/signer', () => ({
@@ -45,33 +21,13 @@ jest.mock('../store/signer', () => ({
 const mockSigner = {} as ethers.providers.JsonRpcSigner;
 
 describe('selectChain', () => {
-  beforeAll(() => {
-    jest
-      .mocked(ethers.providers.Web3Provider.prototype.getSigner)
-      .mockReturnValue(mockSigner);
-  });
-
-  beforeEach(() => {
-    jest.mocked(ethers.providers.Web3Provider.prototype.getSigner).mockClear();
-  });
-
   it('should select chain', async () => {
+    jest.mocked(fetchSigner).mockResolvedValueOnce(mockSigner);
+
     await selectChain(mainnetChain);
 
     expect(switchNetwork).toHaveBeenCalledWith({ chainId: mainnetChain.id });
-    expect(srcChain.set).toHaveBeenCalledWith(mainnetChain);
-    expect(destChain.set).toHaveBeenCalledWith(taikoChain);
-
-    expect(ethers.providers.Web3Provider.prototype.send).toHaveBeenCalledWith(
-      'eth_requestAccounts',
-      undefined,
-    );
-
-    // By default the signer is not updated as
-    // there might be no change in the account.
-    expect(
-      ethers.providers.Web3Provider.prototype.getSigner,
-    ).toHaveBeenCalled();
+    expect(fetchSigner).toHaveBeenCalled();
 
     expect(signer.set).toHaveBeenCalledWith(mockSigner);
